NaCl: Remove a call to BrokerDuplicateHandle() in DidOpenResource().
The call was replaced with IPC::TakePlatformFileForTransit(), which uses
attachment brokering.

In general, Chrome IPC only allows parameters to be read once. This is
particularly relevant for brokered parameters, since reading the parameter also
takes ownership of it. This CL changes the order of the parameters in the IPC
message PpapiHostMsg_OpenResource to move ppapi::proxy::SerializedHandle to the
end. This allows a change in logic in NaClIPCAdapter that prevents the parameter
from being read twice.

BUG=493414

Review URL: https://codereview.chromium.org/1879653002

Cr-Original-Commit-Position: refs/heads/master@{#386534}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bf74d59549913d7f608521917facb59262346b98