Remove all uses of skia::RefPtr and stale includes
This CL is the chrome/browser/, chrome/renderer/ and ppapi/-specific
sections of the change.

Finish deprecating skia::RefPtr in favor of sk_sp<>.
This should significantly reduce the amount of manual reference
management Chromium and Blink authors need to do for Skia objects.

We still have some APIs vending raw pointers to ref-counted Skia
objects which will need to be wrapped in sk_sp<>, but that should
be easier and more predictable. This CL does *not* try to remove
all cases of raw pointers to Skia objects in Chromium.

R=fmalita@chromium.org,pkasting@chromium.org,bbudge@chromium.org,piman@chromium.org
TBR=dfalcantara@chromium.org,rdsmith@chromium.org,rockot@chromium.org
BUG=skia:5077,604716

Review-Url: https://codereview.chromium.org/1949473002
Cr-Original-Commit-Position: refs/heads/master@{#391358}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5e1236bb62070b34cb63af9a8596f26f9d518b34