Generate param traits size methods for IPC files in chrome/ (and traits it depends on).
This allows us to reuse these traits in Mojo if needed.

I didn't add GetSize to every single trait in chrome/ and below; just enough to get the chrome message generator to compile with the size traits.

This is a followup to r393058.

Review-Url: https://codereview.chromium.org/1977643002
Cr-Original-Commit-Position: refs/heads/master@{#393526}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3db6b6d9b24c3c0d44df4929eb8b44605d338cbf