Null check dispatcher for Pepper video decode.
As explained in the added comment, ProvidePictureBuffers is called by the
graphics system in response to a message from the GPU process. These messages
will not be synchronized with the lifetime of the plugin so we need to
null-check the dispatcher before dereferencing.

BUG=477008

Review-Url: https://codereview.chromium.org/2378823003
Cr-Original-Commit-Position: refs/heads/master@{#421652}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d1edc257fd11b931d3a09f937e441e82d846fcde