Fix Android gamepad test crashes and reenable.
These tests were crashing on Android with SIGBUS BUS_ADRALN because ARM
requires 32-bit alignment for accesses to doubles.  The #pragma
pack(push, 1) in the headers broke the alignment.

The pragma was originally added by scottmg@ in
https://bugs.webkit.org/show_bug.cgi?id=78022 "to avoid difference
across compiler settings in different modules/platforms".  There seems
to be no particular reason a packing of 1 was chosen, so switch it to 4
to match WebInputEvent.h's packing.

BUG=147549

Review-Url: https://codereview.chromium.org/2618773003
Cr-Original-Commit-Position: refs/heads/master@{#442422}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8c1ccaf38c8405d554bb6d2bc413aa97c2f3536b