Plumb separate DidScroll message from PDF plugin to embedder.
The change in r509091 added code to notify the embedder whenever the
PDF was scrolled so that touch-selection handles, which are drawn by the
browser and not the PDF viewer, could have their positions adjusted to
match the scroll.

This was done by having PDFiumEngine call OnSelectionChanged() for
each scroll, which triggers re-sending the selection coordinates to the
embedder. However, it seems this also triggers a DCHECK as
OnSelectionChanged() should not be called when the selection is in
a form area.

This CL plumbs an independent pathway from the PDF plugin to the
embedder so that this DCHECK can be avoided.

Bug: 784932
Change-Id: I8db353f0afa0e8c8bd625735d27e00d58c9b0dac
Reviewed-on: https://chromium-review.googlesource.com/782404
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: dsinclair <dsinclair@chromium.org>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Commit-Queue: James MacLean <wjmaclean@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#518936}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 289948d07fd71d92862af6523794d366633e1d80