Fix TouchSelection handles for PDF
The fix for https://crbug.com/784932 seems to have broken the
touch selection handles for PDF; sfter scrolling they're in
the wrong place.

Going back to the original solution, but this time making sure that
on scroll we don't update the selected text, thus avoiding the
DCHECK that was at the root of 784932.

Bug: 829406
Change-Id: I387d120d105c351d80c337ba1646a3cb1eaf40dc
Reviewed-on: https://chromium-review.googlesource.com/998073
Reviewed-by: dsinclair <dsinclair@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Commit-Queue: James MacLean <wjmaclean@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#550398}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 04445ad561d99ba5fd6718e0a5f78a8452da7eda