gpu: Add COMMANDS_ISSUED_TIMESTAMP_CHROMIUM query
This CL adds the GL_COMMANDS_ISSUED_TIMESTAMP_CHROMIUM query type to the
raster and GLES2 command buffers. A client can use this query to get the
timestamp at which preceding commands have been issued to the driver.

The use case is in the follow-up CL: we want to measure the wall time it
takes for a raster task created in the renderer to start being issued to
the driver in the GPU process. This essentially makes sure that there
are no regressions between these events: this is useful to make sure
that the way we process hardware accelerated image decodes does not
cause raster work to systematically start later than with software image
decodes.

This query is similar to the GL_TIMESTAMP query (see the ARB_timer_query
specification). The main differences are:

1) The QueryCounter() behavior of GL_TIMESTAMP is asynchronous, but it
   is undesired because it gives us the timestamp when all preceding
   commands have been fully realized: we don't care about commands being
   completed; we only care that they've been issued to the driver.

2) The GetIntegerv()/GetInteger64v() behavior of GL_TIMESTAMP is
   undesired because it is synchronous.

Bug: 995155
Test: Updated relevant unit tests.
Change-Id: Ibc52b2c530151d2fe0740180a8f643d992ee8b5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1753206
Commit-Queue: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Bill Budge <bbudge@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#689299}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 21e9815207d7ad044b5092d39ed48883a7649551