Add enums for a "fit to paper" scaling print setting
The new enum values (added to Blink and Pepper) will be handled
properly by the PDF plugin, but we want to avoid passing it to any
other plugins, notably Flash. Accordingly, a previous CL [1] exposed
to the UI whether the content being printed is a PDF to prevent the
new scaling option for Flash. Additionally, this CL adds a check in
PepperPluginInstanceImpl::PrintBegin() to avoid sending the new enum
to a non-PDF plugin.

A future CL will implement the functionality of "fit to paper" scaling
and another will add an option to the UI.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1788484

Bug: 989978
Change-Id: I91e821f8d04d1da0fb718b71fb7f331e45dfed07
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1780961
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Raymes Khoury <raymes@chromium.org>
Commit-Queue: Daniel Hosseinian <dhoss@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#698031}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7f2909c48023547e48755aff48e540ed4ad5c0ed